+++ CHANGELOG.md
@@ -5,9 +5,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
-## [0.9.9] - 2021-03-29
+## [0.9.9] - 2021-04-21
+### Added
+- Add basic model args camel snake 
 ### Changed
 - Add ability to use kwargs for model creation
+### Fixed
+- Model Kwargs vs param 
+
 
 ## [0.9.8] - 2021-03-29
 ### Changed